move BaseMapTraits to TraitClasses
